fix: names of variables. List item component moved
